+++ ceammc/ext/doc/docs.cmake
@@ -72,6 +72,7 @@ set(DOC_FILES
     conv.sec2samp
     conv.sec2str
     conv.str2sec
+    counter
     data.fifo
     data.float
     data.dict